+++ lib/src/paged_datatable_column_header.dart
@@ -9,8 +9,9 @@ class _PagedDataTableHeaderRow<TKey extends Object, TResult extends Object>
 
   @override
   Widget build(BuildContext context) {
+    var config = PagedDataTableConfiguration.of(context);
     return SizedBox(
-      height: 56,
+      height: config.columnsHeaderHeight,
       child: Stack(
         fit: StackFit.expand,
         children: [